backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.0.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lxml==5.3.2
MarkupSafe==3.0.2
nltk==3.8.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.2.1
psycopg2-binary==2.9.9
pycparser==2.22
pypdfium2==4.30.1
python-docx==0.8.11
regex==2024.11.6
SQLAlchemy==2.0.40
tqdm==4.67.1
typing_extensions==4.13.2
Werkzeug==3.1.3
boto3==1.26.0
python-dotenv==1.0.0